Run after closure on background queue
`Deferred.Promise.wait(_:)` blocks synchronously.
When `after` targets the main queue, its block then doesn't get a chance
to run till after the `wait` has returned, even if we're waiting
past when the closure gets enqueued.

Enqueuing on a background queue lets us fill the promise
while the main queue is blocked waiting for the promise to be filled,
so the rendezvous now succeeds.